Git: Add alias for get log on merge-base to HEAD
Given two commits A and B, git merge-base A B will output a commit
which is reachable from both A and B through the parent relationship.

For example, with this topology:

```
         o---o---o---B
        /
---o---1---o---o---o---A
```

the merge base between A and B is 1.

`git log-base-forward A` in this case equal `git log 1..A`
`git log-base-forward B` in this case equal `git log 1..B`